+++ src/lib.rs
@@ -229,217 +229,25 @@
 //! <https://docs.rs> builds on nightly, these appear in the published docs even
 //! though stable users cannot use them.
 
-use regex::Regex;
-use std::collections::HashMap;
-use std::ffi::{OsStr, OsString};
-use std::process::Command;
-use std::process::Output;
-use std::sync::LazyLock;
-
 mod cmd_error;
-mod command;
 mod exit_status;
+mod functions;
 mod named_command;
 mod named_output;
 mod which_problem;
 
 pub use cmd_error::CmdError;
 pub use exit_status::ExitStatusFromCode;
+#[cfg(command_resolved_envs)]
+pub use functions::display_env_vars;
+pub use functions::{
+    display, display_with_env_keys, nonzero_captured, nonzero_streamed, on_system_error,
+};
 pub use named_command::{CommandWithName, NamedCommand};
 pub use named_output::NamedOutput;
 pub use named_output::OutputWithName;
 #[cfg(feature = "which_problem")]
 pub use which_problem::map_which_problem;
-
-// https://github.com/jimmycuadra/rust-shellwords/blob/d23b853a850ceec358a4137d5e520b067ddb7abc/src/lib.rs#L23
-static QUOTE_ARG_RE: LazyLock<Regex> =
-    LazyLock::new(|| Regex::new(r"([^A-Za-z0-9_\-.,:/@\n])").expect("clippy checked"));
-
-/// Converts a command and its arguments into a user readable string
-///
-/// # Examples
-///
-/// ```rust
-/// use std::process::Command;
-/// use fun_run;
-///
-/// let name = fun_run::display(Command::new("bundle").arg("install"));
-/// assert_eq!(String::from("bundle install"), name);
-/// ```
-#[must_use]
-pub fn display(command: &mut Command) -> String {
-    vec![command.get_program().to_string_lossy().to_string()]
-        .into_iter()
-        .chain(command.get_args().map(OsStr::to_string_lossy).map(|arg| {
-            if QUOTE_ARG_RE.is_match(&arg) {
-                format!("{arg:?}")
-            } else {
-                format!("{arg}")
-            }
-        }))
-        .collect::<Vec<String>>()
-        .join(" ")
-}
-
-/// Converts a command, and specified environment variables to user readable string
-///
-/// Takes an environment variable **key** and if it will be used when the command runs
-/// prepends the `<key>=<value>` pair to the front of the command.
-///
-/// **Warning:** By default a [`Command`] will inherit environment variables from the parent process.
-/// Limit environment variables to non-sensitive keys or use [`Command::env_clear`] and explicitly
-/// [`Command::envs`] to set only what you need.
-///
-/// This safer alternative to [`display_with_env_keys`] resolves environment variables from
-/// [`Command::get_resolved_envs`]. That function will automatically account for
-/// inherited environment variables and any env modifications such as [`Command::env_clear`]
-/// or [`Command::env_remove`].
-///
-/// **Note:** Requires a nightly toolchain. This function relies on the unstable
-/// [`command_resolved_envs`](https://github.com/rust-lang/rust/issues/149070)
-/// feature, which is auto-detected at build time. On a stable toolchain it does
-/// not exist, so calling it (or referring to it) will not compile.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::process::Command;
-/// use fun_run;
-///
-/// let mut command = Command::new("bundle");
-/// command.arg("install").envs([("RAILS_ENV", "production")]);
-///
-/// let name = fun_run::display_env_vars(&mut command, ["RAILS_ENV"]);
-/// assert_eq!(String::from(r#"RAILS_ENV="production" bundle install"#), name);
-/// ```
-#[cfg(command_resolved_envs)]
-#[must_use]
-pub fn display_env_vars<T, K>(cmd: &mut Command, keys: T) -> String
-where
-    T: IntoIterator<Item = K>,
-    K: Into<OsString>,
-{
-    let env: HashMap<OsString, OsString> = cmd.get_resolved_envs().collect();
-    display_with_env_keys(cmd, env, keys)
-}
-
-/// Converts a command, arguments, and specified environment variables to user readable string
-///
-/// Useful for showing usage of a command that uses environment variables for configuration.
-///
-/// # Examples
-///
-/// ```rust
-/// use std::process::Command;
-/// use fun_run;
-/// use std::collections::HashMap;
-///
-/// let mut env = std::env::vars().collect::<HashMap<_,_>>();
-/// env.insert("RAILS_ENV".to_string(), "production".to_string());
-///
-/// let mut command = Command::new("bundle");
-/// command.arg("install").envs(&env);
-///
-/// let name = fun_run::display_with_env_keys(&mut command, &env, ["RAILS_ENV"]);
-/// assert_eq!(String::from(r#"RAILS_ENV="production" bundle install"#), name);
-/// ```
-///
-/// There's no guarantee that the env provided was passed to construct the Command.
-/// A [`Command`] can also inherit environment variables from the parent.
-///
-/// Note that [`Command::env_clear`] and [`Command::env_remove`] both change the Command's
-/// env var inheritance behavior.
-#[must_use]
-pub fn display_with_env_keys<E, K, V, I, O>(cmd: &mut Command, env: E, keys: I) -> String
-where
-    E: IntoIterator<Item = (K, V)>,
-    K: Into<OsString>,
-    V: Into<OsString>,
-    I: IntoIterator<Item = O>,
-    O: Into<OsString>,
-{
-    display_name_with_env_keys(cmd.name(), env, keys)
-}
-
-fn display_name_with_env_keys<E, K, V, I, O>(name: String, env: E, keys: I) -> String
-where
-    E: IntoIterator<Item = (K, V)>,
-    K: Into<OsString>,
-    V: Into<OsString>,
-    I: IntoIterator<Item = O>,
-    O: Into<OsString>,
-{
-    let env = env
-        .into_iter()
-        .map(|(k, v)| (k.into(), v.into()))
-        .collect::<HashMap<OsString, OsString>>();
-
-    keys.into_iter()
-        .map(|key| {
-            let key = key.into();
-            format!(
-                "{}={:?}",
-                key.to_string_lossy(),
-                env.get(&key).cloned().unwrap_or_else(|| OsString::from(""))
-            )
-        })
-        .chain([name])
-        .collect::<Vec<String>>()
-        .join(" ")
-}
-
-#[cfg(not(any(unix, windows)))]
-compile_error!(
-    "fun_run constructs `ExitStatus` values and only supports `unix` and `windows` targets"
-);
-
-/// Converts a [`std::io::Error`] into a [`CmdError`] which includes the formatted command name
-#[must_use]
-pub fn on_system_error(name: String, error: std::io::Error) -> CmdError {
-    CmdError::SystemError(name, error)
-}
-
-/// Converts an [`Output`] into an error when status is non-zero
-///
-/// When calling a [`Command`] and streaming the output to stdout/stderr
-/// it can be jarring to have the contents emitted again in the error. When this
-/// error is displayed those outputs will not be repeated.
-///
-/// Use when the [`Output`] comes from a source that was already streamed.
-///
-/// To to include the results of stdout/stderr in the display of the error
-/// use [`nonzero_captured`] instead.
-///
-/// # Errors
-///
-/// Returns Err when the [`Output`] status is non-zero
-pub fn nonzero_streamed(name: String, output: impl Into<Output>) -> Result<NamedOutput, CmdError> {
-    let output = output.into();
-    if output.status.success() {
-        Ok(output.named(name))
-    } else {
-        Err(CmdError::NonZeroExitAlreadyStreamed(output.named(name)))
-    }
-}
-
-/// Converts an [`Output`] into an error when status is non-zero
-///
-/// Use when the [`Output`] comes from a source that was not streamed
-/// to stdout/stderr so it will be included in the error display by default.
-///
-/// To avoid double printing stdout/stderr when streaming use [`nonzero_streamed`]
-///
-/// # Errors
-///
-/// Returns Err when the [`Output`] status is non-zero
-pub fn nonzero_captured(name: String, output: impl Into<Output>) -> Result<NamedOutput, CmdError> {
-    let output = output.into();
-    if output.status.success() {
-        Ok(output.named(name))
-    } else {
-        Err(CmdError::NonZeroExitNotStreamed(output.named(name)))
-    }
-}
 
 #[cfg(test)]
 mod tests {
